requirements.txt
streamlit
playwright
beautifulsoup4
pillow